% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_DL_det.R
\name{summarize_DL_det}
\alias{summarize_DL_det}
\title{Summarize newly downloaded detections by transmitter}
\usage{
summarize_DL_det(
  det_dir,
  DL_dates,
  capture_data,
  time.zone,
  Date_format,
  Transmitters,
  tags.to.remove
)
}
\arguments{
\item{det_dir}{A directory path which houses folders of download logs/data named by date (see \code{\link[=sort_VRL_logs]{sort_VRL_logs()}}).}

\item{DL_dates}{An optional vector of download dates of interest. This must match the format of the names
of the date-specific folders in the directory. If not provided, all downloads from the directory will be imported.}

\item{capture_data}{An optional \code{data.frame}, \code{tibble}, or \code{data.table} containing relevant capture information.
\emph{\strong{MUST}} include a column titled \code{"Transmitter"} with the full transmitter names and a column titled \code{"Date.Caught"}.
Note: \code{Date.Caught} does not need to be pre-formatted as a \code{Date} or \code{POSIXct} object.}

\item{time.zone}{Optional. The time zone of the capture_data. Defaults to "US/Eastern".}

\item{Date_format}{Optional. The \code{POSIX} format of the dates of capture. Defaults to \code{"\%m/\%d/\%Y"}.}

\item{Transmitters}{An optional vector of full transmitter names. If capture_data is not supplied, this must be included.}

\item{tags.to.remove}{An optional vector including the full transmitter names of tags wished to be removed from the summary.}
}
\value{
Returns a listed object containing:
\describe{
\item{summary}{A \code{data.table} with columns \code{"n_detections"}, \code{"n_stations"}, \code{"first_det"}, and \code{"last_det"}}
\item{detections}{A \code{data.table} with all detections associated with the specified \code{Transmitter} names}
}
}
\description{
Summarizes newly downloaded detection data which have been sorted into date-specific folders using
\code{\link[=sort_VRL_logs]{sort_VRL_logs()}}. You must provide either a data object of capture data (\code{data.frame}, \code{tibble}, or \code{data.table})
or a vector of complete transmitter names.
}
\details{
\itemize{
\item If providing \code{capture_data}, it \emph{\strong{MUST}} have a column named \code{"Transmitter"} populated with complete
transmitter names (e.g., "A69-9001-57441") as well as a column named \code{"Date.Caught"} populated by the date
in which the animal was caught.
\item The \code{capture_data} object can have an optional column titled "Spp.Name" populated with the scientific names
of the fish caught (e.g., "Pristis pectinata", "P. pectinata"). If present, the output data summary will be
ordered alphabetically by species. If "pectinata" is present in the name, the order will have P. pectinata
at the top, followed by the remaining species in alphabetical order.
\item If \code{capture_data} is provided, all columns in \code{capture_data} are joined with the output \code{summary}.
}
}
\examples{
\dontrun{
setwd("G:/Charlotte Harbor Common/Sawfish/Acoustic Database Backups")
root_dir <- "2025 Receiver Logs"
det_dir <- "Logs by Date"

# Import tag information
fish_ids <- data.table::fread("CSV Files/Acoustic Tagged Fish.csv")

# Dates of interest
dates <- c("2025-02-28", "2025-05-16", "2025-09-26")

# Tags correspond to known deceased fish
tags.to.remove <- c("A69-9001-54949", "A69-9001-60330", "A69-9001-54939")

# Summarize and view data
det_summary <- summarize_DL_det(
  det_dir = det_dir,
  DL_dates = dates,
  capture_data = fish_ids,
  tags.to.remove = tags.to.remove
)

# Optional reorder across additional columns
data.table::setorder(det_summary$summary, Spp.Name, ProjName, sex, Transmitter)

# Save a csv
write.csv(det_summary$summary, row.names = FALSE,
          file = paste0(root_dir, "/detection_summary_", Sys.Date(), ".csv"))
}

}
\seealso{
\code{\link[=sort_VRL_logs]{sort_VRL_logs()}}, \code{\link[=load_detections_by_date]{load_detections_by_date()}}
}
